chore(storage): Extract ech interface into it own file
