+++ src/Uno.Foundation/IAsyncActionInternal.cs
@@ -1,0 +1,13 @@
+#nullable enable
+
+using System;
+using System.Linq;
+using System.Threading.Tasks;
+
+namespace Windows.Foundation
+{
+	internal interface IAsyncActionInternal : IAsyncAction
+	{
+		Task Task { get; }
+	}
+}
