+++ src/Uno.Foundation/IAsyncOperation.cs
@@ -1,5 +1,3 @@
-using System.Threading.Tasks;
-
 namespace Windows.Foundation
 {
 	public partial interface IAsyncOperation<TResult> : IAsyncInfo
@@ -7,10 +5,5 @@ namespace Windows.Foundation
 		AsyncOperationCompletedHandler<TResult> Completed { get; set; }
 
 		TResult GetResults();
-	}
-
-	internal interface IAsyncOperationInternal<TResult> : IAsyncOperation<TResult>
-	{
-		Task<TResult> Task { get; }
 	}
 }
