+++ src/Uno.Foundation/IAsyncOperationInternal.cs
@@ -1,0 +1,13 @@
+#nullable enable
+
+using System;
+using System.Linq;
+using System.Threading.Tasks;
+
+namespace Windows.Foundation
+{
+	internal interface IAsyncOperationInternal<TResult> : IAsyncOperation<TResult>
+	{
+		Task<TResult> Task { get; }
+	}
+}
