+++ src/Uno.Foundation/IAsyncOperationWithProgress.cs
@@ -2,7 +2,6 @@
 using System.Collections.Generic;
 using System.Runtime.InteropServices;
 using System.Text;
-using System.Threading.Tasks;
 using Windows.Foundation.Metadata;
 
 namespace Windows.Foundation
@@ -12,11 +11,6 @@ namespace Windows.Foundation
 
 	public partial interface IAsyncOperationWithProgress<TResult, TProgress> : IAsyncInfo
 	{
-	}
-
-	internal interface IAsyncOperationWithProgressInternal<TResult, TProgress> : IAsyncOperationWithProgress<TResult, TProgress>
-	{
-		Task<TResult> Task { get; }
 	}
 }
